inv-driver: named opcode consts + per-family encoder colocation + SO_PEERCRED gate
Codec gains a single canonical home for every byte literal that crosses
a wire boundary: opcodes.go (L1/L2 sentinels, family-specific cmd bytes,
sub-codes, ACK codes), info.go (uint8 model-code consts), limits.go
(MinPanelLimitW / MaxPanelLimitW exported for the firmware-imposed
20..500 W per-panel gate). The EncodeSetPower dispatcher and the
cmd/inv-driver/setpower CLI now reference these names instead of raw
hex literals; model codes collapse to a single uint8 set instead of
shadowing uint32 + uint8 forms. Encoders for each family colocate with
their decoder — EncodeSetPowerDS3 in ds3.go, EncodeSetPowerQS1A in
qs1a.go (renamed from EncodeSetPowerDSPNew), EncodeSetPowerC3 in the
new yc600.go. The setpower.go dispatcher is the only file that names
all three families.

Test-only fixtures move to a sibling codectest package so the
production binary no longer carries hex.DecodeString init or the data
strings.

Ingestor.HandleWithPeer plumbs a verified peer UID from the UDS layer
through to the downstream routing gate. ipc.Server reads SO_PEERCRED
on each connection (Linux build tag; non-Linux returns -1, gate skipped
on dev hosts). routeDownstream now requires the verified UID to be in
the configured ControllerUIDs set in addition to the Hello-declared
backend allow list; the Hello.Backend string is informational, the UID
is authoritative. cmd/inv-driver/main.go exposes -controller-backends
and -controller-uids flags with defaults inv-driver-cli / "0".

handleTelemetry rejects Panels and LifetimeRaw arrays larger than 16
entries so a malicious publisher cannot inflate the per-channel SQLite
writes.

+++ codec/info_test.go
@@ -119,7 +119,7 @@ func TestDecodeInfoReply_BadCmdBytes(t *testing.T) {
 func TestPhaseFromModel(t *testing.T) {
 	t.Parallel()
 	cases := []struct {
-		model uint32
+		model uint8
 		want  uint32
 	}{
 		{7, 1},
